xproto: normal dependency of libxau
libxau uses headers of xproto and also requires xproto in its .pc file.
Here, we make xproto a normal dependency of libxau. As a result,
packages that depend on libxau and libxcb (that depend on libxau) do not
need to require xproto as the latter will be "brought" in automatically.

Signed-off-by: Maxim Belkin <maxim.belkin@gmail.com>