Multiple formulae: remove redundant dependencies
1. 'libxcb' has a normal dependency on 'libpthread-stubs'. Therefore,
any formula that depends on 'libxcb' does not need to specify the latter
dependency as it will be added/installed automatically.

2. 'libx11' depends on 'libxau' (through 'libxcb') that a normal
dependency on 'xproto'. 'libx11' also has a normal dependency on
'kbproto' (in its '.pc' file). Therefore, anything that depends on
'libx11' does not need to specify neither 'kbproto' no 'xproto' as their
dependencies as they will be added automatically

Signed-off-by: Maxim Belkin <maxim.belkin@gmail.com>